pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Andrey Yurpalov <a.yurpalov@mail.ru>"]
readme = "README.md"
packages = [{include = "gendiff"}]


[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"


[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
